moved tokenblocklist to separate file
