+++ project/config/__init__.py
@@ -1,2 +1,2 @@
-from .application_factory import create_app, DB, TokenBlocklist
+from .application_factory import create_app, DB, JWT
 from .errors import register_error_handlers
