+++ project/models/__init__.py
@@ -1,4 +1,5 @@
 from .crypto import CryptoModel
 from .stock import StockModel
 from .user import UserModel
+from .token import TokenBlocklist
 from .utils import delete_from_database, add_to_database
